- (prolang.y)   add_string_constant(): Added size to be allocated to an error message   Corrected a bunch of format specifiers   Added a missing line break in an error message in add_lvalue_code()   epilog(): changed size from int to p_int.   Changed some comparisons between signed and unsigned types in (lines     12229, 12322, 12935). - (lex.c)   parse_numeric_escape(): Removed a comparison of unsigned long < 0 which is     never true and in this case not needed. - (i-eval_cost.h)   add_eval_cost_n(): changed a comparison between signed and unsigned types
git-svn-id: svn://svn.bearnip.com/ldmud/trunk@2420 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/i-eval_cost.h
@@ -39,7 +39,7 @@ add_eval_cost_n (int32 added_cost, uint repetitions)
             return MY_TRUE;
 
         if (max_eval_cost - eval_cost < added_cost
-         || (max_eval_cost - eval_cost) / repetitions < added_cost
+         || (int32)((max_eval_cost - eval_cost) / repetitions) < added_cost
            )
         {
             total_evalcost += max_eval_cost - eval_cost + 1;
